Improve error messages on add new server dialog
Closes #438

Note: this solution is different than what @jasonblais suggested.
@jasonblais suggested not highlighting the field that hasn't been
filled out yet, but I think it's a better UX to continue to highlight
them all as it currently does, but to clarify the error messages. Also
notably, as far as I can tell, the react-bootstrap [forms
API](https://react-bootstrap.github.io/components/forms/) does not
appear to support selectively highlighting some fields in the form but
not others anyway.