[GH 2874] Fix category for deb and clairfy directory information in README for linux (#2884)
* Remove expliced category from deb

The expliced category value for deb is not correct. The linux section
has a correct category value and deb is part of its target. So the
category from linux should be applied if there is no expliced one in the
deb section.
Issue: mattermost/desktop#2874

* Linux README: clarify where to find directory

Change the example directory to the path, where it is at a deb
installation. Add a hint that a reader in a local README already
found the correct directory.
Issue: mattermost/desktop#2874